Fix logic error where a non-existing file would cause issues deleting a cache.

+++ EssentialNostrTests/EventCache/CodableEventStoreTests.swift
@@ -47,6 +47,9 @@ class CodableEventStore {
     }
 
     func deleteCachedEvents(completion: @escaping EventStore.DeletionCompletion) {
+        guard FileManager.default.fileExists(atPath: storeURL.path()) else {
+            return completion(nil)
+        }
         do {
             try FileManager.default.removeItem(at: storeURL)
             completion(nil)
